+++ web/src/routes/+layout.ts
@@ -1,0 +1,6 @@
+import { checkAppVersion } from '$lib/utils/check-app-version';
+import { LayoutLoad } from '.svelte-kit/types/src/routes/$types';
+
+export const load: LayoutLoad = async ({ url }) => {
+	return { url };
+};
