+++ web/src/routes/auth/logout/+server.ts
@@ -1,3 +1,4 @@
+import { json } from '@sveltejs/kit';
 import { api, serverApi } from '@api';
 import type { RequestHandler } from '@sveltejs/kit';
 
@@ -5,15 +6,14 @@ export const POST: RequestHandler = async () => {
 	api.removeAccessToken();
 	serverApi.removeAccessToken();
 
-	return {
+	return json({
+		ok: true
+	}, {
 		headers: {
 			'Set-Cookie': [
 				'immich_is_authenticated=deleted; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT;',
 				'immich_access_token=delete; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT'
 			]
-		},
-		body: {
-			ok: true
 		}
-	};
+	});
 };
