+++ web/src/routes/auth/register/+page.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+	import AdminRegistrationForm from '$lib/components/forms/admin-registration-form.svelte';
+	import { api } from '@api';
+</script>
+
+<svelte:head>
+	<title>Admin Registration - Immich</title>
+</svelte:head>
+
+<section class="h-screen w-screen flex place-items-center place-content-center">
+	<AdminRegistrationForm />
+</section>
