+++ web/src/routes/data/album/get-all-albums/+server.ts
@@ -1,3 +1,4 @@
+import { json } from '@sveltejs/kit';
 import { AlbumResponseDto, serverApi } from '@api';
 import type { RequestEvent, RequestHandler, RequestHandlerOutput } from '@sveltejs/kit';
 
@@ -7,12 +8,13 @@ export const GET = async ({
 	try {
 		const isShared = url.searchParams.get('isShared') === 'true' || undefined;
 		const { data } = await serverApi.albumApi.getAllAlbums(isShared);
+		throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+		// Suggestion (check for correctness before using):
+		// return json(data);
 		return {
 			body: data
 		};
 	} catch {
-		return {
-			status: 500
-		};
+		return new Response(undefined, { status: 500 });
 	}
 };
