+++ web/src/routes/data/user/get-my-user-info/+server.ts
@@ -1,0 +1,17 @@
+import { json } from '@sveltejs/kit';
+import { serverApi, UserResponseDto } from '@api';
+import type { RequestHandlerOutput } from '@sveltejs/kit';
+
+export const GET = async (): Promise<RequestHandlerOutput<UserResponseDto>> => {
+	try {
+		const { data } = await serverApi.userApi.getMyUserInfo();
+		throw new Error("@migration task: Migrate this return statement (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+		// Suggestion (check for correctness before using):
+		// return json(data);
+		return {
+			body: data
+		};
+	} catch {
+		return new Response(undefined, { status: 500 });
+	}
+};
