+++ web/src/routes/photos/[assetId]/+page.ts
@@ -1,0 +1,14 @@
+import { redirect } from '@sveltejs/kit';
+export const prerender = false;
+
+import { browser } from '$app/env';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({}) => {
+	//!TODO refactor session
+	// if (!browser && !session.user) {
+	// 	throw redirect(302, '/auth/login');
+	// } else {
+	// 	throw redirect(302, '/photos');
+	// }
+};
